날씨 서비스 형식 일부 수정 (2)
192c5a457c4ef1f2c81d1218ad2e0a139f7320d9

+++ app/src/main/java/com/teamhub/notiget/model/weather/openweather/HourlyModel.java
@@ -9,7 +9,7 @@ public class HourlyModel {
     public long dt;
     public double feels_like;
     public int humidity;
-    public int pop;
+    public double pop;
     public int pressure;
     public double temp;
     public int visibility;
